feat: identify/separate AVX-specific classes
